feat: Add NATS microservice for order communication
This commit adds the necessary code changes to integrate NATS as a microservice for order communication. It includes the addition of the "nats" package as a dependency in the package.json file and the creation of the "NatsModule" in the "transports" directory. The "NatsModule" is imported in both the "app.module.ts" and "orders.module.ts" files to enable the use of NATS in the application.